Added support for release events.
I originally looked into adding support for ReleaseEvents in Octocrab, but releases have *so* much information that I didn't want to spend the time to investigate all the structures involved. Through a series of unfortunate events, I've basically removed Octocrab and just implemented minimal structures.

This allows for a little bit more flexibility in how we process payloads, as we only need to extract the information we care about.